requirements.txt
alembic==1.15.2
bcrypt==4.0.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.3.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
requests==2.32.3
concurrent-log-handler==0.9.25
cryptography==3.4.7
google-genai==1.31.0
